srcs/django/requirements.txt
Django>=3.2,<4.0
psycopg2-binary
requests
django-environ
daphne
redis
# channels==4.0.0
channels_redis
djangorestframework
Pillow
neat-python
